+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Relics & Rituals chat card improvements
 - Chat card styling improvements
 - Fix a bug when an actor was updating too often
-- The options for Devil's Luck is outputted into the chat when clicking the Fevil's Luck label in the character sheet
+- The options for Devil's Luck is outputted into the chat when clicking the Devil's Luck label in the character sheet
 - Hide Mystical Mishaps for players
 - Add dice sound to chat card roll buttons